feat: add un-integrated config package (#129)
<!--
Thank you for your Pull Request. Please provide a description above and review
the requirements below. Bug fixes and new features should include tests.
-->

## Motivation
Gas assets work stream + deposit screener feature will need to use broader `NocturneConfig` to store info. This PR precedes both work streams.
<!--
Explain the context and why you're making that change. What is the problem
you're trying to solve? In some cases there is not a problem and this can be
thought of as being the motivation for your change.
-->

## Solution
- Add `@nocturne-xyz/config` package which includes `NocturneConfig` object
- Move proxy types out of `deploy` into `config`
- NOTE: this is not integrated anywhere yet, integration into SDK will occur in next PR
<!--
Summarize the solution and provide any necessary context needed to understand
the code change.
-->

## Proof

<!--
If features/changes is hard to test e2e, include a video or image proving you've
tested your solution and it works.
-->

## PR Checklist

- [x] Added Tests
- [ ] Updated Documentation (N/A)
- [x] Updated CHANGELOG.md for the appropriate package
- [ ] Tested site with snap (we haven't automated this yet) (N/A)
- [ ] Re-built & tested circuits if any of them changed (N/A)