[SDK] Try to make even number of joinsplits ## Motivation
- We currently use dummy notes if there is odd number of joinsplits (JSs)
- Though it doesn’t cost anymore to spend 2 notes than 1 in single JS, thus we should always attempt to join the 2nd note to consolidate user balances into fewer big notes rather than many small notes
- Should always try to spend 2 notes to reduce dust in sdk `gatherNotes`

## Solution

After running our usual algorithm, whenever there are an odd number of notes
selected, we select the next smallest note to add to the joinsplit.

We may still need to pad with a dummy when all notes are in use and there is
an odd number of them.

Existing tests work and I added one test and updated one test case to better
cover the new code.
## Proof

Tests working well, ran on my computer also and deposit -> transfer flow worked.