[GEO] introduce new middleware and apply it (#491)
## Motivation

We have to keep track of various access patterns to our services. 

## Solution

Now that we have AWS WAF tagging some of our incoming requests with new headers we can react to that information. As a starting proof of concept, let's just log the new stuff.

## Proof

Will test in a deployed environment

## PR Checklist

- [ ] added tests
- [ ] updated documentation
- [ ] added changeset if necessary
- [x] tested in dev/testnet
- [x] tested site with snap (we haven't automated this yet)
- [x] re-built & tested circuits if any of them changed
- [x] updated contracts storage layout (if contracts were updated)