[postgres] add local image and migrations (#501)
## Motivation

For many features we will want a queryable backend. Geo-tracking in particular
has created this desire in this case but I strongly suspect postgres will be used
to persist other information as well.

## Solution

If we want to add postgres to our stack the first step is to responsibly set up the users and migrations. The first step in that is the local development experience with an eye toward the deployed experience.

* use docker-compose to stand up and configure a postgres database instance
* use Flyway to execute migrations against our postgres database locally
* use prettier to format our SQL migrations (optional)

Long term, we will likely use flyway and docker to run the migrations with 
different parameters in our remote environments.

## Proof

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/v1lR6ZTcsV4NqRqfLNgi/d6f3fa91-2319-4138-ab2d-1c6c46766d16/image.png)