patch(build): yarn build should build e2e tests (#526)
## Motivation

This was a sharp corner on my last PR. I made some updates that changed some
signatures. I ran `yarn build` at the root of the repo to ensure that I had
patched all of the callsites. The expected behavior was that it would build
everything in the monorepo but it did not.
## Solution

I understand that we're not using the artifacts but `yarn build` at the
monorepo level should catch any typescript errors. This adds a basic
`tsc` call that will _check_ the e2e tests. This has the added  benefit
of running on the pre-commit hook which should prevent any unexpected
CI failures and save us some CI time.

The downside is that it does take a few seconds to run on local and in CI.
## Proof

In CI it will run.