feat(trigger-migrations): on push to main ## Motivation
We currently automatically create a new image and tag it on push to the
main branch. Now, we want to automatically trigger the migrations
after the push of the flyway image.

## Solution

Use the AWS EventBridge that we set up to trigger the migrations to run
after we upload the new image with the `latest` tag.