feat(local-kms): start adding kms signature code ## Motivation
We currently use in memory keys for the Attestation key. While these
keys are injected via an encrypted channel it is still not an ideal
scenario. HSMs provide the highest level of security for private key
management. In the AWS ecosystem, KMS provides a very cheap layer on
top of HSMs with access managed by IAM roles.

Importantly, the key material can never leave the HSMs. This is a step
in the direction of using HSMs to sign transactions inside nocturne
services.

## Solution

Introduce local-kms into our docker-compose file. This will allow us to
interact with KMS using the same tools that we will use in AWS but for
local dev.

We will likely allow in memory keys for backward compatibility and
simplicity in local environments.

## Proof

WIP

## PR Checklist

- [n/a] added tests
- [n/a] updated documentation
- [n/a] added changeset if necessary
- [n/a] tested in dev/testnet
- [n/a] tested site with snap (we haven't automated this yet)
- [n/a] re-built & tested circuits if any of them changed
- [n/a] updated contracts storage layout (if contracts were updated)