feat(defender): Add http timeout to Defender ## Motivation
We are seeing infinite hang on our test actor and we believe the culprit
is an infinite timeout on our http calls to Defender.

## Solution

Configure defender to use an http agent with timeout set. This also
introduces keepAlive to reduce tcp/ip chatter.

## Proof

WIP will likely have to deploy this one
## PR Checklist

- [n/a] added tests
- [n/a] updated documentation
- [] added changeset if necessary
- [wip] tested in dev/testnet
- [WIP] tested site with snap (we haven't automated this yet)
- [n/a] re-built & tested circuits if any of them changed
- [n/a] updated contracts storage layout (if contracts were updated)